requirements.txt
fastapi~=0.103.1
fastapi-utils~=0.2.1
yfinance~=0.2.36
uvicorn[standard]~=0.23.2
pandas_datareader~=0.10.0
pytrends~=4.9.2
GoogleNews~=1.6.14
transformers~=4.35.0

torch~=2.0.1
torchvision~=0.15.2
spacy~=3.6.1
pika~=1.3.2